requirements.txt
notebooklm-py>=0.3.0
httpx>=0.27.0
